+++ light-desk/CHANGELOG.md
@@ -25,8 +25,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Introduce new color constant `COLOR_RGB_BLACK`
 - Fixed bugs that would occur when a child added to a group multiple times
 - [breaking]: start using Color utilities from `@synesthesia-project/compositor`
-- [breaking]: rework backend interfaces to unify defining and mutating props
-  and adding / removing listeners
+- [breaking]: rework backend interfaces to unify defining and mutating props,
+  adding / removing listeners, and child-parent relationships.
 
 ## [2.1.1] - 2022-01-11